chore(suite-native): cleanup navigation helpers
