+++ scripts/circular-dependencies/madge-snapshot.txt
@@ -250,5 +250,5 @@ suite-native/device/src/index.ts > suite-native/device/src/components/Bootloader
 suite-native/device/src/index.ts > suite-native/device/src/components/BootloaderModeScreen.tsx > suite-native/device-manager/src/index.ts > suite-native/device-manager/src/components/DeviceManager.tsx > suite-native/device-manager/src/components/DeviceSwitch.tsx > suite-native/device-manager/src/components/DeviceSwitchContent.tsx
 suite-native/device/src/index.ts > suite-native/device/src/components/BootloaderModeScreen.tsx > suite-native/device/src/selectors.ts > suite-native/firmware/src/index.ts > suite-native/firmware/src/components/FirmwareInstallationScreenContent.tsx
 suite-native/navigation/src/components/ScreenHeader.tsx > suite-native/navigation/src/components/ScreenHeaderContent.tsx
-suite-native/navigation/src/index.ts > suite-native/navigation/src/useNavigateToInitialScreen.ts
+suite-native/navigation/src/index.ts > suite-native/navigation/src/hooks/useNavigateToInitialScreen.ts
 suite-native/react-native-graph/src/LineGraphProps.ts > suite-native/react-native-graph/src/CreateGraphPath.ts
