+++ suite-native/navigation/src/hooks/useNavigateToInitialScreen.ts
@@ -2,7 +2,7 @@ import { useCallback } from 'react';
 
 import { ParamListBase, useNavigation } from '@react-navigation/native';
 
-import { StackToTabCompositeProps } from './index';
+import { StackToTabCompositeProps } from '../index';
 
 type NavigationProp = StackToTabCompositeProps<ParamListBase, string, ParamListBase>;
