+++ suite-native/navigation/src/index.ts
@@ -2,11 +2,10 @@ export * from './navigators';
 export * from './routes';
 export * from './types';
 export * from './config';
-export * from './useInterceptNativeNavigation';
-export * from './useNavigateToInitialScreen';
+export * from './hooks/useInterceptNativeNavigation';
+export * from './hooks/useNavigateToInitialScreen';
 export * from './useScrollDivider';
 export * from './routeUtils';
-export * from './hooks/useNavigationRoute';
 export * from './hooks/useOverrideBackNavigation';
 export * from './components/TabBar';
 export * from './components/Screen';
@@ -17,4 +16,5 @@ export * from './components/ScreenFooterGradient';
 export * from './components/NavigationContainerWithAnalytics';
 export * from './components/GoBackIcon';
 export * from './components/ScrollViewContext';
+export * from './hooks/useNavigationRouteMatch';
 export * from './hooks/useDisableIOSGesture';
